requirements.txt
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.0.3
grpcio==1.70.0
grpcio-tools==1.70.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
protobuf==5.29.4
py4j==0.10.9.7
pyspark==3.5.5
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
